move class to features
